fix: remove circular imports and add a linting step for this
