+++ package.json
@@ -29,9 +29,10 @@
     "build": "tsc",
     "dev": "tsc-watch --project tsconfig.json --onSuccess 'node --inspect .'",
     "debug": "tsc-watch --project tsconfig.json --onSuccess 'node --inspect-brk .'",
-    "lint": "npm run lint:commit && npm run lint:eslint",
+    "lint": "npm run lint:commit && npm run lint:eslint && npm run lint:circular",
     "lint:eslint": "eslint \"src/**/*.ts\" && (cd test && eslint \"**/*.ts\")",
     "lint:commit": "commitlint --from=$(git rev-parse origin/staging)",
+    "lint:circular": "fadge detect-cycles --allow-ignores=true \"./{src,test}/**/*.ts\"",
     "format": "prettier --write '{src,test}/**/*.{js,ts,json,yml}'",
     "format:check": "prettier --check '{src,test}/**/*.{js,ts,json,yml}'"
   },
@@ -67,6 +68,7 @@
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^3.4.1",
+    "fadge": "^0.0.1",
     "jest": "^26.6.3",
     "nock": "^13.0.11",
     "prettier": "^2.4.1",
